requirements.txt
pyyaml>=6.0.1
streamlit>=1.37.0
